backend/requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
Flask-Login==0.6.2
Flask-SQLAlchemy==3.1.1
Flask-Cors==4.0.0
mysqlclient==2.2.0
python-dotenv==1.0.0
PyJWT==2.8.0
pytest==7.4.2
pytest-cov==4.1.0
black==23.9.1
flake8==6.1.0
schedule==1.2.0